Turn class into mixin. Update readme. Remove index.js. Update main and module in package.json
