+++ computing-lit-element.js
@@ -1,1 +1,5 @@
-export { ComputingLitElement, html } from './src/ComputingLitElement.js';
+import { computingMixin } from './src/computingMixin.js';
+import { LitElement, html } from 'lit-element';
+
+const ComputingLitElement = computingMixin(LitElement);
+export { ComputingLitElement, html, computingMixin };
