--- index.js
@@ -1,1 +1,0 @@
-export { ComputingLitElement } from './src/ComputingLitElement.js';
