+++ package.json
@@ -9,8 +9,8 @@
     "type": "git",
     "url": "git+https://github.com/Neovici/computing-lit-element.git"
   },
-  "main": "index.js",
-  "module": "index.js",
+  "main": "computing-lit-element.js",
+  "module": "computing-lit-element.js",
   "scripts": {
     "start": "es-dev-server --app-index demo/index.html --node-resolve --open --watch",
     "start:compatibility": "es-dev-server --app-index demo/index.html --compatibility all --node-resolve --open --watch",
